Use special connection factory instead of inheritance.
