+++ simpleclient_pushgateway/src/main/java/io/prometheus/client/exporter/BasicAuthHttpConnectionFactory.java
@@ -1,0 +1,36 @@
+package io.prometheus.client.exporter;
+
+import java.io.IOException;
+import java.io.UnsupportedEncodingException;
+import java.net.HttpURLConnection;
+
+public class BasicAuthHttpConnectionFactory implements HttpConnectionFactory {
+    private final HttpConnectionFactory originConnectionFactory;
+    private final String basicAuthHeader;
+
+    public BasicAuthHttpConnectionFactory(HttpConnectionFactory connectionFactory, String user, String password) {
+        this.originConnectionFactory = connectionFactory;
+        this.basicAuthHeader = encode(user, password);
+    }
+
+    public BasicAuthHttpConnectionFactory(String user, String password) {
+        this(new DefaultHttpConnectionFactory(), user, password);
+    }
+
+    @Override
+    public HttpURLConnection create(String url) throws IOException {
+        HttpURLConnection connection = originConnectionFactory.create(url);
+        connection.setRequestProperty("Authorization", basicAuthHeader);
+        return connection;
+    }
+
+    private String encode(String user, String password) {
+        try {
+            byte[] message = (user + ":" + password).getBytes("UTF-8");
+            String encoded = javax.xml.bind.DatatypeConverter.printBase64Binary(message);
+            return String.format("Basic %s", encoded);
+        } catch (UnsupportedEncodingException e) {
+            throw new IllegalArgumentException(e);
+        }
+    }
+}
