+++ simpleclient_pushgateway/src/main/java/io/prometheus/client/exporter/DefaultHttpConnectionFactory.java
@@ -1,0 +1,12 @@
+package io.prometheus.client.exporter;
+
+import java.io.IOException;
+import java.net.HttpURLConnection;
+import java.net.URL;
+
+public class DefaultHttpConnectionFactory implements HttpConnectionFactory {
+    @Override
+    public HttpURLConnection create(String url) throws IOException {
+        return (HttpURLConnection) new URL(url).openConnection();
+    }
+}
