+++ simpleclient_pushgateway/src/main/java/io/prometheus/client/exporter/HttpConnectionFactory.java
@@ -1,0 +1,8 @@
+package io.prometheus.client.exporter;
+
+import java.io.IOException;
+import java.net.HttpURLConnection;
+
+public interface HttpConnectionFactory {
+    HttpURLConnection create(String url) throws IOException;
+}
